+++ mamaearth/src/components/BabyPage/BabyPage.jsx
@@ -1,0 +1,7 @@
+import { Products } from "../Page/Products";
+
+export const BabyPage = () =>{
+    return(
+        <Products url={"https://mama-earth.herokuapp.com/baby_page"} poster={"https://images.ctfassets.net/66mrrren2unf/1LH7b5JOWtuTiarFdJ2QNA/07434c4bf1d6e104ada7e8c7f13f57d3/desktop-bathing-.jpg?q=40"}/>
+    )
+}
