--- mamaearth/src/components/BabyPage/card.css
@@ -1,56 +1,0 @@
-.cards {
-  border: 1px solid lightgray;
-  width: 270px;
-  height: 490px;
-  margin: auto;
-  text-align: center;
-}
-.rating {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 5px;
-}
-.rating p {
-  border-right: 1px solid grey;
-  padding-right: 10px;
-}
-.rate-review {
-  display: flex;
-  justify-content: center;
-  min-height: 24px;
-}
-.review {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 5px;
-  padding-left: 10px;
-}
-.desc {
-  font-size: 1rem;
-  font-weight: 400;
-}
-.price {
-  font-size: 1.2rem;
-  font-weight: bold;
-  margin-bottom: 5px;
-}
-.cart_btn{
-    background-color: white;
-    border: 1px solid rgb(0, 175, 239);
-    color: rgb(0, 175, 239);
-    font-weight: bold;
-    border-radius: 5px;
-    text-align: center;
-    width: 90%;
-    height: 50px;
-}
-.item-img{
-    width: 270px;
-    height: 300px;
-}
-.cart_btn:hover{
-  background-color: rgb(0, 175, 239);
-  color: white;
-}
